Code refactored for better export
