+++ test/arithmetic/divide.test.js
@@ -1,4 +1,4 @@
-import divide from '../../src/arithmetic/divide';
+import { divide } from '../../src';
 
 test('Should be able to divide two integers', () => {
   expect(divide(4, 2)).toBe(2);
