+++ test/range/max.test.js
@@ -1,4 +1,4 @@
-import max from '../../src/range/max';
+import { max } from '../../src';
 
 test('Should give max entry from a given array', () => {
   expect(max([1, 3, 4])).toBe(4);
